move config to separate module
