+++ src/registry/proxy/mod.rs
@@ -1,7 +1,5 @@
 mod proxy_config;
 mod remote_image;
 
-pub use proxy_config::{
-    DownloadRemoteImageError, RegistryProxiesConfig, SingleRegistryProxyConfig,
-};
+pub use proxy_config::DownloadRemoteImageError;
 pub use remote_image::RemoteImage;
